+++ src/backend/resolvers/queries/index.js
@@ -9,7 +9,6 @@ const organizations = require('./organizations')
 const qualityOfAccessScores = require('./qualityOfAccessScores')
 const collections = require('./collections')
 const newTreatmentPlans = require('./newTreatmentPlans')
-const emailUsers = require('./emailUsers')
 const testEmailGroups = require('./testEmailGroups')
 
 module.exports = {
@@ -23,7 +22,6 @@ module.exports = {
   regimens,
   qualityOfAccessScores,
   newTreatmentPlans,
-  emailUsers,
   ...organizations,
   testEmailGroups,
 }